Factor out getSubjectColor function from Dashboard
Make Dashboard.vue less cluttered.
